Implement new command run-locally.
The main advantage is that it supports executing interactive commands,
including runme-in-runme, and pipes (|).

On top of that, there is the new package internal/command. It gives
a better interface to create commands, as well as unifies the idea
of a program config that now is defined in a proto file.