Use PATH from session to look for programs (#555)
If `PATH` is present in the session, use it to look for the program
paths. This is especially important for situation when within a session
a virtual env is created and consecutive code blocks should be executed
within it.

Fixes https://github.com/stateful/runme/issues/552

Port this change to https://github.com/stateful/runme/pull/548.