Define AI GRPC service to support using AI to generate cells (#573)
We'd like to use AI to generate cells. In particular, we'd like to
support using [Foyle](foyle.io) to add AI capabilities.

The initial plan is to use a gRPC service to allow RunMe to communicate
with Foyle (or potentially other AI services). This proto needs to
define that gRPC service.