move document package from internal to pkg (#588)
The purpose of this PR is to make it possible to reuse runme's Markdown
-> Notebook code.
In particular I'd like to use it to parse ChatGPT output into a notebook
so that code blocks can be rendered in RunMe notebooks.

In Foyle, there is currently a bug with parsing out code blocks nested
inside lists jlewi/foyle#114. Runme's code properly handles this use
case. This will allow us to reuse RunMe's code to fix this problem.