Implement fifo based shell env collector (#617)
Replace the shell env collector from a file-based to named pipe-based
(fifo). For Windows, it will be a named pipe, but it will be done in a
follow-up PR. 

Also, the env dump regardless of the destination is encrypted. 
The encryption is mostly to avoid leaking secrets by accidents. Our
threat model assumes execution of runme by the same user without any
additional isolation.

---------

Co-authored-by: Sebastian Tiedtke <sebastiantiedtke@gmail.com>