Allow parser to skip ahead in runes of varying byte-width (#754)
The parser didn't allow the skip ahead in characters because a rune's
multi-byte width was tracked per single iteration. Skipping ahead is
essential for frontmatter detection. Instead of keeping track of the
last rune's width, the specific rune's now passed around. This makes the
parser more stateless.

One downside is likely sacrificing performance for this flexibility.
However, it's expected on the margins. Let's consider benchmarking this
down the road.

However, the fix for
https://github.com/stateful/vscode-runme/issues/1978 is more important.

Signed-off-by: Sebastian (Tiedtke) Huckleberry <sebastiantiedtke@gmail.com>