Unlock DirEnv Support (#757)
This PR introduces support for `.envrc` files at the root of a project.
The root is usually the git repo's root. Runme requires `direnv` to be
installed locally.

**Important: Runme's direnv implementation does not support abritrary
directories and changes between them since it's not a freewheeling
terminal.**

PR includes:
- [x] runnerv1 supports `direnv`
- [x] runnerv1 tests pass in CI/CD

All of runnerv2 will be handled in a separate PR:
- runnerv2 supports `direnv`
- runnerv1 tests pass in CI/CD

---------

Signed-off-by: Sebastian (Tiedtke) Huckleberry <sebastiantiedtke@gmail.com>